Add layout inheritance and new flash messages
