+++ twitter_clone/__init__.py
@@ -1,3 +1,1 @@
-from flask import Flask
 
-app = Flask(__name__)
