+++ twitter_clone/static/js/tweet_clone.js
@@ -1,0 +1,5 @@
+$(document).ready(function() {
+    setTimeout(function() {
+        $(".flashes").fadeOut("slow");
+    }, 4000);
+})
