+++ twitter_clone/templates/static_templates/layout_login.html
@@ -1,0 +1,20 @@
+{% extends "static_templates/layout.html" %}
+{% block content %}
+    <div class="row">
+        <div class="col-md-2"></div>
+        <div class="col-md-8">
+            <form role="form" action="" method="POST">
+    		    <div class="form-group">
+        		    <label for="username">Username</label>
+        		    <input class="form-control" name="username" placeholder="Your username">
+    		    </div>
+    		    <div class="form-group">
+        		    <label for="password">Password</label>
+        		    <input type="password" class="form-control" name="password" placeholder="Your password">
+    		    </div>
+    		    <button type="submit" class="btn btn-default">Submit</button>
+    		</form>
+        </div>
+        <div class="col-md-2"></div>
+    </div>
+{% endblock %}
