+++ twitter_clone/templates/static_templates/layout_profile.html
@@ -1,0 +1,28 @@
+{% extends "static_templates/layout.html" %}
+{% block content %}
+    <div class="row">
+        <div class="col-md-2"></div>
+        <div class="col-md-8">
+            <form role="form" action="" method="POST">
+    		    <div class="form-group">
+        		    <label for="username">Username</label>
+        		    <input class="form-control" name="username" placeholder="Your username" value="{{username}}" readonly>
+    		    </div>
+    		    <div class="form-group">
+        		    <label for="first_name">First name</label>
+        		    <input class="form-control" name="first_name" placeholder="Your first name" value="{{first_name}}">
+    		    </div>
+    		    <div class="form-group">
+        		    <label for="last_name">Last name</label>
+        		    <input class="form-control" name="last_name" placeholder="Your last name" value="{{last_name}}">
+    		    </div>
+    		    <div class="form-group">
+        		    <label for="birth_date">Birth date</label>
+        		    <input class="form-control" name="birth_date" placeholder="Your birth date" value="{{birth_date}}">
+    		    </div>
+    		    <button type="submit" class="btn btn-default">Submit</button>
+    		</form>
+        </div>
+        <div class="col-md-2"></div>
+    </div>
+{% endblock %}
