--- twitter_clone/templates/static_templates/other_feed.html
@@ -1,74 +1,0 @@
-
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>Twitter clone</title>
-
-        <!-- styles -->
-        <link href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" rel="stylesheet">
-        <link href="https://maxcdn.bootstrapcdn.com/font-awesome/4.5.0/css/font-awesome.min.css" rel="stylesheet">
-        <link rel="stylesheet" href="../../static/css/styles.css">
-    </head>
-    <body>
-        <div class="container body-container">
-
-            <nav class="navbar navbar-default navbar-fixed-top">
-  <div class="container">
-    <div class="navbar-header">
-      <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#navbar" aria-expanded="false" aria-controls="navbar">
-        <span class="sr-only">Toggle navigation</span>
-        <span class="icon-bar"></span>
-        <span class="icon-bar"></span>
-        <span class="icon-bar"></span>
-      </button>
-      <a class="navbar-brand" href="/">Twitter clone</a>
-    </div>
-    <div id="navbar" class="navbar-collapse collapse">
-      <ul class="nav navbar-nav navbar-right">
-
-            <li><a href="/profile"><i class="fa fa-user" aria-hidden="true"></i>{{me}}</a></li>
-            {% if session.logged_in %}
-            <li><a href="{{ url_for('logout') }}"><i class="fa fa-sign-out" aria-hidden="true"></i> Log out</a></li>
-            {% else %}
-            <li><a href="{{ url_for('login') }}"><i class="fa fa-sign-out" aria-hidden="true"></i> Log in</a></li>
-            {% endif %}
-
-
-      </ul>
-    </div>
-  </div>
-</nav>
-
-
-<div class="row">
-    <div class="col-sm-2"></div>
-    <div class="col-sm-8">
-
-
-        <div class="row">
-
-            {% for index, tweet, timestamp in tweets %}
-            <div class="well well-large">
-
-
-                <div>
-                    <strong>@{{username}}</strong>
-                    <span class="label label-primary">{{timestamp}}</span>
-                </div>
-                <div>{{tweet}}</div>
-
-
-            </div>
-            {% endfor %}
-
-        </div>
-    </div>
-    <div class="col-sm-2"></div>
-</div>
-
-        </div>
-    </body>
-
-    <!-- scripts -->
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js" type="text/javascript"></script>
-</html>
