--- twitter_clone/templates/static_templates/profile.html
@@ -1,67 +1,0 @@
-
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>Twitter clone</title>
-
-        <!-- styles -->
-        <link href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" rel="stylesheet">
-        <link href="https://maxcdn.bootstrapcdn.com/font-awesome/4.5.0/css/font-awesome.min.css" rel="stylesheet">
-        <link rel="stylesheet" href="../../static/css/styles.css">
-    </head>
-    <body>
-        <div class="container body-container">
-
-            <nav class="navbar navbar-default navbar-fixed-top">
-  <div class="container">
-    <div class="navbar-header">
-      <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#navbar" aria-expanded="false" aria-controls="navbar">
-        <span class="sr-only">Toggle navigation</span>
-        <span class="icon-bar"></span>
-        <span class="icon-bar"></span>
-        <span class="icon-bar"></span>
-      </button>
-      <a class="navbar-brand" href="/">Twitter clone</a>
-    </div>
-    <div id="navbar" class="navbar-collapse collapse">
-      <ul class="nav navbar-nav navbar-right">
-
-            <li><a href="/profile"><i class="fa fa-user" aria-hidden="true"></i>{{username}}</a></li>
-            <li><a href="/logout?next=%2Fprofile"><i class="fa fa-sign-out" aria-hidden="true"></i> Log out</a></li>
-
-      </ul>
-    </div>
-  </div>
-</nav>
-
-
-<div class="row">
-    <div class="col-md-12">
-        <form role="form" action="" method="POST">
-		  <div class="form-group">
-		    <label for="username">Username</label>
-		    <input class="form-control" name="username" placeholder="Your username" value="{{username}}" readonly>
-		  </div>
-		  <div class="form-group">
-		    <label for="first_name">First name</label>
-		    <input class="form-control" name="first_name" placeholder="Your first name" value="{{first_name}}">
-		  </div>
-		  <div class="form-group">
-		    <label for="last_name">Last name</label>
-		    <input class="form-control" name="last_name" placeholder="Your last name" value="{{last_name}}">
-		  </div>
-		  <div class="form-group">
-		    <label for="birth_date">Birth date</label>
-		    <input class="form-control" name="birth_date" placeholder="Your birth date" value="{{birth_date}}">
-		  </div>
-		  <button type="submit" class="btn btn-default">Submit</button>
-		</form>
-    </div>
-</div>
-
-        </div>
-    </body>
-
-    <!-- scripts -->
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js" type="text/javascript"></script>
-</html>
